backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
psycopg[binary]==3.2.6
psycopg_pool==3.2.6
python-multipart==0.0.20
Pillow==11.2.1
cloudinary==1.41.0
